* moved prev logged users list to account store from app store
